requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
sentence-transformers>=2.2.2

# Web Interface
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pypdf2>=3.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# API & Networking
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.4.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.7.0
cachetools>=5.3.0
